Clean up docs
