+++ Sources/PostgresKit/Deprecations/PostgresConnectionSource+PostgresConfiguration.swift
@@ -1,6 +1,5 @@
 import NIOSSL
 import Atomics
-import AsyncKit
 import Logging
 import PostgresNIO
 import NIOCore
