--- Sources/PostgresKit/Docs.docc/index.md
@@ -1,3 +1,0 @@
-# ``PostgresKit``
-
-PostgresKit is a library to provide a simple Swift interface to PostgresNIO using SQLKit.
