+++ Sources/PostgresKit/PostgresKit.docc/PostgresKit.md
@@ -1,0 +1,28 @@
+# ``PostgresKit``
+
+@Metadata {
+    @TitleHeading(Package)
+}
+
+PostgresKit is a library providing an SQLKit driver for PostgresNIO.
+
+## Overview
+
+This package provides the "foundational" level of support for using [Fluent] with PostgreSQL by implementing the requirements of an [SQLKit] driver. It is responsible for:
+
+- Managing the underlying PostgreSQL library ([PostgresNIO]),
+- Providing a two-way bridge between PostgresNIO and SQLKit's generic data and metadata formats,
+- Presenting an interface for establishing, managing, and interacting with database connections.
+
+> Note: The FluentKit driver for PostgreSQL is provided by the [FluentPostgresDriver] package.
+
+## Version Support
+
+This package uses [PostgresNIO] for all underlying database interactions. It is compatible with all versions of PostgreSQL and all platforms supported by that package.
+
+> Important: There is one exception to the above at the time of this writing: This package requires Swift 5.7 or newer, whereas PostgresNIO continues to support Swift 5.6.
+
+[SQLKit]: https://swiftpackageindex.com/vapor/sql-kit
+[PostgresNIO]: https://swiftpackageindex.com/vapor/postgres-nio
+[Fluent]: https://swiftpackageindex.com/vapor/fluent-kit
+[FluentPostgresDriver]: https://swiftpackageindex.com/vapor/fluent-postgres-driver
