+++ Tests/PostgresKitTests/PostgresKitTests.swift
@@ -261,12 +261,3 @@ extension Bar: PostgresNonThrowingEncodable, PostgresArrayEncodable, PostgresDec
     static var psqlFormat: PostgresFormat { .binary }
     static var psqlArrayType: PostgresDataType { .int8Array }
 }
-
-let isLoggingConfigured: Bool = {
-    LoggingSystem.bootstrap { label in
-        var handler = StreamLogHandler.standardOutput(label: label)
-        handler.logLevel = env("LOG_LEVEL").flatMap { Logger.Level(rawValue: $0) } ?? .info
-        return handler
-    }
-    return true
-}()
